requirements.txt
PyQt6>=6.4.0
PyQtWebEngine>=6.4.0
requests>=2.28.0
PySocks>=1.7.1
cryptography>=3.4.8